Separate out ibc contract/msg code into own files
