+++ contracts/ibc-reflect-send/src/ibc_msg.rs
@@ -1,0 +1,37 @@
+#![allow(clippy::field_reassign_with_default)] // see https://github.com/CosmWasm/cosmwasm/issues/685
+
+use cosmwasm_std::{Coin, ContractResult, CosmosMsg, HumanAddr};
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+/// This is the message we send over the IBC channel
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum PacketMsg {
+    Dispatch { msgs: Vec<CosmosMsg> },
+    WhoAmI {},
+    Balances {},
+}
+
+/// All IBC acknowledgements are wrapped in `ContractResult`.
+/// The success value depends on the PacketMsg variant.
+pub type AcknowledgementMsg<T> = ContractResult<T>;
+
+/// This is the success response we send on ack for PacketMsg::Dispatch.
+/// Just acknowledge success or error
+pub type DispatchResponse = ();
+
+/// This is the success response we send on ack for PacketMsg::WhoAmI.
+/// Return the caller's account address on the remote chain
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct WhoAmIResponse {
+    pub account: HumanAddr,
+}
+
+/// This is the success response we send on ack for PacketMsg::Balance.
+/// Just acknowledge success or error
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct BalancesResponse {
+    pub account: HumanAddr,
+    pub balances: Vec<Coin>,
+}
