+++ contracts/ibc-reflect-send/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod contract;
+pub mod ibc;
+pub mod ibc_msg;
 pub mod msg;
 pub mod state;
