+++ contracts/ibc-reflect-send/src/msg.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::field_reassign_with_default)] // see https://github.com/CosmWasm/cosmwasm/issues/685
 
-use cosmwasm_std::{Coin, ContractResult, CosmosMsg, Empty, HumanAddr};
+use cosmwasm_std::{Coin, CosmosMsg, Empty, HumanAddr};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
@@ -100,35 +100,4 @@ impl From<AccountData> for AccountResponse {
             remote_balance: input.remote_balance,
         }
     }
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum PacketMsg {
-    Dispatch { msgs: Vec<CosmosMsg> },
-    WhoAmI {},
-    Balances {},
-}
-
-/// All acknowledgements are wrapped in `ContractResult`.
-/// The success value depends on the PacketMsg variant.
-pub type AcknowledgementMsg<T> = ContractResult<T>;
-
-/// This is the success response we send on ack for PacketMsg::Dispatch.
-/// Just acknowledge success or error
-pub type DispatchResponse = ();
-
-/// This is the success response we send on ack for PacketMsg::WhoAmI.
-/// Return the caller's account address on the remote chain
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct WhoAmIResponse {
-    pub account: HumanAddr,
-}
-
-/// This is the success response we send on ack for PacketMsg::Balance.
-/// Just acknowledge success or error
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct BalancesResponse {
-    pub account: HumanAddr,
-    pub balances: Vec<Coin>,
 }
